python_project_template/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.25.0
yfinance>=0.2.0
xgboost>=1.6.0
lightgbm>=3.3.0
ta>=0.10.0
statsmodels>=0.13.0
click>=8.0.0
tqdm>=4.60.0
google-generativeai>=0.3.0